helper_scripts/obd2-tools/requirements.txt
obd==0.7.3
python-can==4.6.1
